% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genChangeData.R
\name{genChangeData}
\alias{genChangeData}
\title{Function to impose an change impact effect}
\usage{
genChangeData(
  pct.change,
  model,
  data,
  panels = NULL,
  eventsite.bnd = NULL,
  noneventcells = NULL
)
}
\arguments{
\item{pct.change}{percentage change to occur. e.g. 50\% gives 50\% site wide decline.  If a redistribution effect (one of \code{eventsite.bnd} or \code{noneventcells} must be specified) then the change refers only to the event site cells.  For an additional site-wide change when using a re-distribution, \code{pct.change} is a vector with the first number the event site re-distribution and the second, the site wide change.}

\item{model}{A glm or gamMRSea model object.}

\item{data}{Data frame used to fit \code{model}}

\item{panels}{Character vector denoting the column of \code{data} containing the panel structure.}

\item{eventsite.bnd}{A data frame containing the coordinates of a polygon defining the region of the defined event The variable names must match the coordinate system in the data.}

\item{noneventcells}{logical (0/1) vector of length the same as the data indicating which cells the event did not occur (1).}
}
\value{
Returns a data frame twice the size of the original input with additional columns for panel id, eventphase and truth.  The truth column represents the input data for the first half (pre-event; eventphase==0) and the second half has the change imposed (post-event; eventphase==1).
}
\description{
Function to impose an change impact effect
}
\examples{
data(nystedA_slim)

initialModel<-MRSea::gamMRSea(response ~ 1 + as.factor(yearmonth)+depth +
                     x.pos + y.pos + offset(log(area)),  data=nysted,
                     family=quasipoisson)

# generate a 20\% site-wide change:
changedata1<-genChangeData(pct.change=80, model=initialModel, data=nysted, panels="panelid")

# generate a 50\% decline in the windfarm footprint:
data("nysted.bndwf")
changedata2<-genChangeData(pct.change=50, model=initialModel, data=nysted, panels="panelid",
                            eventsite.bnd=nysted.bndwf)

}
\author{
Lindesay Scott-Hayward, University of St Andrews
}
