% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateIC.R
\name{generateIC}
\alias{generateIC}
\title{Function to add correlated noise to data}
\usage{
generateIC(data, corrs, panels, newdata, nsim, dots = FALSE)
}
\arguments{
\item{data}{data to be assessed for correlation}

\item{corrs}{matrix definig the correlation for each panel (each row of matrix) defined in the data}

\item{panels}{character giving the column of \code{data} contaning the information about panel structure}

\item{newdata}{simulated sets of data (independent)}

\item{nsim}{number of simulations}

\item{dots}{(default = FALSE). Logical flag stating whether to show progress dots printed in workspace.}
}
\description{
Function to add correlated noise to data
}
\details{
This function is based on work by Iman and Conover (1982).
}
\examples{
data(nystedA_slim)
nysted$panelid<-as.numeric(nysted$unique.transect.label)

model<-glm(response ~ depth  + as.factor(phase),
              family='poisson', data=nysted)

simData<-generateNoise(n=500, response=fitted(model), family='poisson')
corrs<-getCorrelationMat(panel = nysted$panelid, data=nysted$response, dots = FALSE)
generateIC(nysted, corrs, "panelid", simData, nsim=10, dots = FALSE)

}
\author{
LAS Scott-Hayward, University of St Andrews
}
\references{
Iman and Conover (1982).
}

