% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCorrelationMat.R
\name{getCorrelationMat}
\alias{getCorrelationMat}
\title{Function to get the correlation for each panel defined in the data}
\usage{
getCorrelationMat(panel, data, dots = FALSE)
}
\arguments{
\item{panel}{vector identifying the panel to which each row of data belongs.  Data are considered correlated within a panel and independent between panels.}

\item{data}{vector of response data.}
}
\description{
Function to get the correlation for each panel defined in the data
}

