% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact.coverage.R
\name{impact.coverage}
\alias{impact.coverage}
\title{Get the coverage for the impact coefficient}
\usage{
impact.coverage(truebeta, betacis)
}
\description{
@param truebeta true coefficient value
@param betacis dataframe of upper and lower intervals, where the number of rows is the number of bootstrap simulations
}
\details{
@export

@author Lindesay Scott-Hayward, university of St Andrews
}
