% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.raster.R
\name{make.raster}
\alias{make.raster}
\title{Function to make non-regular data into a raster to calculate the differences.}
\usage{
make.raster(ncell.y, xyzdata, z.name = "response")
}
\arguments{
\item{legend.name}{}
}
\description{
Function to make non-regular data into a raster to calculate the differences.
}

