% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeToyData.R
\name{makeToyData}
\alias{makeToyData}
\title{Function to create a toy dataset for package testing}
\usage{
makeToyData(n, changecoef.link = 0, length.panels = 1, binary = FALSE, b0 = 1)
}
\arguments{
\item{n}{number denoting the amount of data to generate.}

\item{changecoef.link}{Coefficient on the scale of the link function to determine the overall change in the population.}

\item{length.panels}{vector denoting the length of the panels to be created.}

\item{binary}{logical (default = FALSE) denoting whether to generate binary data.  If FALSE, poisson data is generated.}

\item{b0}{intercept parameter}
}
\value{
Returns a data frame of columns
\code{x} x covariate
\code{evph} covariate denoting pre(0)/post(1) event
\code{mu} poisson response
\code{mu.bin} binary response (only if \code{binary==TRUE})
\code{panels} column denoting which panel each row corresponds to.  Used to subsequently add correlation to the data.
}
\description{
Function to create a toy dataset for package testing
}
\author{
Lindesay Scott-Hayward, University of St Andrews
}
