% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.diffs.R
\name{plot.diffs}
\alias{plot.diffs}
\title{Function to plot the output from powerSim}
\usage{
\method{plot}{diffs}(power.object, returndata = FALSE, cellsize = NULL)
}
\arguments{
\item{power.object}{power analysis object of class gamMRSea.power}

\item{returndata}{logical stating whether to return the plotting data}

\item{cellsize}{vector of length two denoting the size of the gridcells (width, height).  If unspecified, geom_raster is used for plotting rather than geom_tile.}
}
\value{
Figure showing the mean (middle), lower 2.5\% (left) and upper 97.5\% (right) of estimated differences between before and after the event. (difference = post - pre)
}
\description{
Function to plot the output from powerSim
}
\examples{

# Load two power objects that were created with a 20\% sitewide decline
# 50 simulations were run
data(nysted.power.oc)
data(nysted.power.oc.null)

plot.diffs(nysted.power.oc)

}
\author{
LAS Scott-Hayward
}
