% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacheckfuncs.R
\name{plotMean}
\alias{plotMean}
\title{Plotting function to assess the mean of generated data}
\usage{
plotMean(model, simdat, bins = 15, quants = c(0.025, 0.975))
}
\arguments{
\item{model}{model object of class glm, gam or gamMRSea}

\item{simdat}{matrix of simulated response data.  Each column is a new simulated data set}

\item{bins}{number of bins to be plotted on the histogram}

\item{quants}{vector of length 2 stating the quantiles for the confidence interval bands for the simulated data}
}
\description{
Plotting function to assess the mean of generated data
}
\details{
The histogram returned shows the lower 2.5 and upper 97.5 quantiles (dark grey, long dash lines) for the mean.  The grey solid line is the mean of the fitted values from the inputed  model. The black short dashed line is the mean of the response data.
}
\examples{
data(nystedA_slim)

initialModel<-MRSea::gamMRSea(response ~ 1 + as.factor(yearmonth)+depth +
                     x.pos + y.pos + offset(log(area)),  data=nysted,
                     family=quasipoisson)

nsim<-550
d<-as.numeric(summary(initialModel)$dispersion)
newdat<-generateNoise(nsim, fitted(initialModel), family='poisson', d=d)

plotMean(initialModel, newdat)
}
\author{
Lindesay Scott-Hayward (University of St Andrews)
}
