% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerPlot.R
\name{powerPlot}
\alias{powerPlot}
\title{Function to show how the power to detect change varies with the error rate.}
\usage{
powerPlot(power.object)
}
\arguments{
\item{power.object}{An object of class gamMRSea.power}
}
\value{
Plot showing how the power to detect change varies with increasing error rate. The first grey dashed line is at 1\% and the second at 5\%, traditionally values used as p-value cutoffs. The blue dashed lines indicate the error rate required to get a power of 80\%. This value is given in the title.
}
\description{
Function to show how the power to detect change varies with the error rate.
}
\examples{

# Load two power objects that were created with a 20\% sitewide decline
# 50 simulations were run
data(nysted.power.oc)
data(nysted.power.oc.null)

powerPlot(nysted.power.oc)

}
\author{
LAS Scott-Hayward
}
