% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerSimParallel.R
\name{powerSimPll}
\alias{powerSimPll}
\title{Function to run a power analysis for a model of type 'gamMRSea'}
\usage{
powerSimPll(
  newdat,
  model,
  empdistribution,
  nsim,
  powercoefid,
  predictionGrid = NULL,
  g2k = NULL,
  splineParams = NULL,
  n.boot = 500,
  nCores = 1
)
}
\arguments{
\item{newdat}{matrix of values whose rows are observations and columns are simulated sets of response data}

\item{model}{model object of class \code{gamMRSea}.  This is the model used to generate the data, updated for the inclusion of an event term.}

\item{empdistribution}{Empirical distribution for the runs test}

\item{nsim}{number of simulations to compute}

\item{powercoefid}{the ID of the coefficient in \code{model} that represents the event term}

\item{predictionGrid}{dataframe containing a coordinate grid with covariates.  This grid is used in the assessment of before and after change differences.}

\item{g2k}{object required for \code{MRSea} model with spline terms}

\item{splineParams}{object required for \code{MRSea} model with spline terms}

\item{n.boot}{Number of bootstraps to use to get percentile based confidence intervals for predictions.}

\item{nCores}{How many cores to use for parallel processing.  If one, then parallel computing is not used and a progress bar is printed. If parallel, estimated time till finish along with a progress bar is provided.}
}
\value{
An object of class \code{gamMRSea.power}.

\item{rawrob}{Vector of zeros and ones indicating whether raw or robust s.e. have been used.}
\item{imppvals}{p-values for the event change term for each simulated run}
\item{betacis}{Confidence intervals for the event term coefficient (only if site-wide change)}
\item{powsimfits}{Fitted values each model}
\item{bootdifferences}{data frame of bootstrapped differences (before and after event). Number of rows is the number of rows in the prediction grid divided by 2}
\item{bootpreds}{data frame of bootstrapped predictions.}
\item{Abundance}{Table of sitewide abundance (with upper and lower 95\\\% CI)}
\item{link}{List of the link function (link) and inverse link function (linkinv)}
}
\description{
Function to run a power analysis for a model of type 'gamMRSea'
}
\examples{
# See the vignette for an example of how to run this code.

}
\author{
LAS Scott-Hayward, University of St Andrews
}
