% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact.coverage.R
\name{impact.coverage}
\alias{impact.coverage}
\title{Get the coverage for the impact coefficient}
\usage{
impact.coverage(truebeta, betacis)
}
\description{
@param truebeta thing
 @param betacis thing
}
\details{
@export

@author Lindesay Scott-Hayward
}

