% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.coverage.R
\name{plot.coverage}
\alias{plot.coverage}
\title{Function to plot the confidence intervals from the change parameter}
\usage{
\method{plot}{coverage}(power.object)
}
\arguments{
\item{power.object}{A power analysis object of class gamMRSea.power}
}
\description{
Function to plot the confidence intervals from the change parameter
}
\examples{

# Load a power object that was created with a 20\\\% sitewide decline
# 50 simulations were run
data(nysted.power.oc)

plot.coverage(nysted.power.oc)

}
\author{
LAS Scott-Hayward, University of St Andrews
}

