% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.diffs.R
\name{plot.diffs}
\alias{plot.diffs}
\title{Function to plot the output from powerSim}
\usage{
\method{plot}{diffs}(power.object, returndata = FALSE)
}
\description{
Function to plot the output from powerSim
}
\author{
Lindesay Scott-Hayward (University of St Andrews)
}

