% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.preds.R
\name{plot.preds}
\alias{plot.preds}
\title{Function plot the mean, lower 2.5\% and upper 97.5\% of predicted animal counts before and after the event.}
\usage{
\method{plot}{preds}(power.object, returndata = FALSE, cellsize)
}
\arguments{
\item{power.object}{power analysis object of class gamMRSea.power}

\item{returndata}{logical stating whether to return the plotting data}

\item{cellsize}{vector of length two denoting the size of the gridcells (width, height).  If unspecified, geom_raster is used for plotting rather than geom_tile.}
}
\description{
Function plot the mean, lower 2.5\% and upper 97.5\% of predicted animal counts before and after the event.
}
\examples{

# Load two power objects that were created with a 20\\\% sitewide decline
# 50 simulations were run
data(nysted.power.oc)
data(nysted.power.oc.null)

plot.preds(nysted.power.oc)

}
\author{
LAS Scott-Hayward
}

