% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.sigdiff.R
\name{plot.sigdiff}
\alias{plot.sigdiff}
\title{Function to plot the output from powerSim}
\usage{
\method{plot}{sigdiff}(powerout, coordinates, tailed = "two",
  error.rate = 0.05, adjustment = "none")
}
\description{
Function to plot the output from powerSim
}

