% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacheckfuncs.R
\name{plotMean}
\alias{plotMean}
\title{Plotting function to assess the mean of generated data}
\usage{
plotMean(model, simdat, bins = 15, quants = c(0.025, 0.975))
}
\arguments{
\item{model}{model object of class glm, gam or gamMRSea}

\item{simdat}{matrix of simulated response data.  Each column is a new simulated data set}

\item{bins}{number of bins to be plotted on the histogram}

\item{quants}{vector of length 2 stating the quantiles for the confidence interval bands for the simulated data}
}
\description{
Plotting function to assess the mean of generated data
}
\author{
Lindesay Scott-Hayward (University of St Andrews)
}

