% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacheckfuncs.R
\name{plotVariance}
\alias{plotVariance}
\title{Plotting function to assess the mean-variance relationship of generated data}
\usage{
plotVariance(model, simdat, cuts = 20, quants = c(0.025, 0.975))
}
\arguments{
\item{model}{model object of class glm, gam or gamMRSea}

\item{simdat}{matrix of simulated response data.  Each column is a new simulated data set}

\item{cuts}{The number of cut points for calculation of residual variance}

\item{quants}{vector of length 2 stating the quantiles for the confidence interval bands for the simulated mean-variance relationship}
}
\description{
Plotting function to assess the mean-variance relationship of generated data
}
\author{
Lindesay Scott-Hayward (University of St Andrews)
}

