% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval.coverage.null.R
\name{pval.coverage.null}
\alias{pval.coverage.null}
\title{Function to estimate the coverage of the overall change impact coefficient.}
\usage{
pval.coverage.null(newdat.ind, newdat.corr = NULL, model, nsim, powercoefid,
  empdistnull = NULL)
}
\arguments{
\item{newdat.ind}{simulated independent response data}

\item{model}{glm or gamMRSea model object}

\item{nsim}{number of simulations in newdat.ind}

\item{powercoefid}{coefficient id that relates to the eventchange term}

\item{empdistnull}{if known, provide the empirical distribution for the runs test.}
}
\description{
Function to estimate the coverage of the overall change impact coefficient.
}
\author{
Lindesay Scott-Hayward
}

