% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval.coverage.null.R
\name{pval.coverage.null}
\alias{pval.coverage.null}
\title{Function to estimate the coverage of the overall change impact coefficient.}
\usage{
pval.coverage.null(newdat.ind, newdat.corr = NULL, model, nsim, powercoefid,
  splineParams = NULL, empdistnull = NULL)
}
\arguments{
\item{model}{thing}

\item{nsim}{thing}

\item{powercoefid}{thing}

\item{newdat}{thing}

\item{g2k}{thing}
}
\description{
Function to estimate the coverage of the overall change impact coefficient.
}
\author{
Lindesay Scott-Hayward
}

