% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval.differences.R
\name{pval.differences}
\alias{pval.differences}
\title{Function to get p-values for individual cell differences}
\usage{
pval.differences(null, estimate, family = FALSE)
}
\arguments{
\item{null}{thing}

\item{estimate}{thing}
}
\description{
Function to get p-values for individual cell differences
}
\examples{

}

