% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.MRSea.power.R
\name{summary.gamMRSea.power}
\alias{summary.gamMRSea.power}
\title{Summary function for objects of class gamMRSea.power}
\usage{
\method{summary}{gamMRSea.power}(power.object, null.object = NULL, truebeta)
}
\arguments{
\item{power.object}{An object of class gamMRSea.power}

\item{thing}{}
}
\description{
Summary function for objects of class gamMRSea.power
}

