% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.MRSea.power.R
\name{summary.gamMRSea.power}
\alias{summary.gamMRSea.power}
\title{Summary function for objects of class gamMRSea.power}
\usage{
\method{summary}{gamMRSea.power}(power.object, null.object = NULL,
  truebeta = NULL)
}
\arguments{
\item{power.object}{An object of class gamMRSea.power}

\item{null.object}{Object containing information regarding the error rate when no change is present in the data.}

\item{truebeta}{Value indicating the proportion of data affected by the change. Only required for a site wide change.}
}
\value{
printed summary output

If only the power.object is provided, then the summary is reduced or informs the reader of some unknowns.
}
\description{
Summary function for objects of class gamMRSea.power
}
\examples{

# Load two power objects that were created with a 20\% sitewide decline
# 50 simulations were run
data(nysted.power.oc)
data(nysted.power.oc.null)

summary(nysted.power.oc, nysted.power.oc.null, truebeta=log(0.8))

}
\author{
LAS Scott-Hayward
}
